requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
-e git+https://github.com/unclecode/crawl4ai.git@897e0173618d20fea5d8952ccdbcdad0febc0fee#egg=Crawl4AI
cryptography==45.0.2
cssselect==1.3.0
cssselect2==0.8.0
distro==1.9.0
fake-http-header==0.3.5
fake-useragent==2.2.0
filelock==3.18.0
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
h11==0.16.0
html2text==2025.4.15
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
humanize==4.12.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
litellm==1.70.2
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
nltk==3.9.1
numpy==2.2.6
openai==1.75.0
packaging==25.0
pillow==10.4.0
playwright==1.52.0
propcache==0.3.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pydyf==0.11.0
pyee==13.0.0
Pygments==2.19.1
pyOpenSSL==25.1.0
pyperclip==1.9.0
pyphen==0.17.2
python-dotenv==1.1.0
PyYAML==6.0.2
rank-bm25==0.2.2
readability-lxml==0.8.4.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.7
tf-playwright-stealth==1.1.2
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
zopfli==0.2.3.post1
